Moved a bunch of files out, and fixed up a few issues
