+++ src/ReactGraph/ExpressionDefinition.cs
@@ -1,0 +1,22 @@
+using System.Linq.Expressions;
+using ReactGraph.Construction;
+using ReactGraph.NodeInfo;
+
+namespace ReactGraph
+{
+    public class ExpressionDefinition<T> : IDefinitionIdentity
+    {
+        public ExpressionDefinition(Expression expression, NodeType nodeType, string nodeName)
+        {
+            NodeType = nodeType;
+            NodeName = nodeName;
+            Path = ExpressionStringBuilder.ToString(expression);
+            Root = ExpressionParser.GetRootOf(expression);
+        }
+
+        public object Root { get; private set; }
+        public string Path { get; private set; }
+        public string NodeName { get; private set; }
+        public NodeType NodeType { get; private set; }
+    }
+}
