+++ src/ReactGraph/IDefinitionIdentity.cs
@@ -1,0 +1,15 @@
+using ReactGraph.NodeInfo;
+
+namespace ReactGraph
+{
+    public interface IDefinitionIdentity
+    {
+        object Root { get; }
+
+        string Path { get; }
+
+        string NodeName { get; }
+
+        NodeType NodeType { get; }
+    }
+}
