+++ src/ReactGraph/ITargetDefinition.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace ReactGraph
+{
+    public interface ITargetDefinition<in T> : IDefinitionIdentity
+    {
+        Action<T> CreateSetValueDelegate();
+    }
+}
