+++ src/ReactGraph/NodeInfo/ITakeValue`1.cs
@@ -4,6 +4,6 @@ namespace ReactGraph.NodeInfo
 {
     interface ITakeValue<T> : INodeInfo
     {
-        void SetSource(IValueSource<T> formulaNode, Action<Exception> errorHandler);
+        void SetSource(IValueSource<T> sourceNode, Action<Exception> errorHandler);
     }
 }
